Separate Sanbase into three different top-level supervisor trees
